📦 Provide each implementation as a file
* Provide a dedicated file for the following environment: node, browser and nashorn
* Provide a UMD as a fallback
