+++ lib/asciidoctor/js/opal_ext/nashorn/file.rb
@@ -1,16 +1,13 @@
 class File
 
-  def self.file? path
+  def self.read(path)
     %x(
+      var Paths = Java.type('java.nio.file.Paths');
       var Files = Java.type('java.nio.file.Files');
-      return Files.exists(path) && Files.isRegularFile(path);
-    )
-  end
-
-  def self.readable? path
-    %x(
-      var Files = Java.type('java.nio.file.Files');
-      return Files.exists(path) && Files.isReadable(path);
+      var lines = Files.readAllLines(Paths.get(path), Java.type('java.nio.charset.StandardCharsets').UTF_8);
+      var data = [];
+      lines.forEach(function(line) { data.push(line); });
+      return data.join("\n");
     )
   end
