+++ lib/asciidoctor/js/opal_ext/phantomjs/file.rb
@@ -1,0 +1,7 @@
+class File
+
+  def self.read(path)
+    %x(return require('fs').read(path);)
+  end
+
+end
