+++ lib/asciidoctor/js/opal_ext/spidermonkey/file.rb
@@ -1,0 +1,7 @@
+class File
+
+  def self.read(path)
+    %x(return read(path);)
+  end
+
+end
