--- lib/asciidoctor/opal_ext.rb
@@ -1,4 +1,0 @@
-# For backward compatibility with Asciidoctor <= 1.5.5
-require 'asciidoctor/js'
-require 'asciidoctor/converter'
-require 'asciidoctor/js/postscript'
