+++ npm/.eslintrc
@@ -1,0 +1,3 @@
+env:
+  node: true
+  es6: true
