--- npm/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-    "env": {
-        "node": true,
-        "es6": true
-    }
-};
